Fix freespace counter requiring float instead of integer.

+++ sources/app/lib/human.php
@@ -7,7 +7,7 @@ abstract class Human
 	const FACTORS = ['', 'k', 'M', 'G', 'T', 'P', 'E', 'Z', 'Y'];
 	const UNIT_SIZES = ['B', 'kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
 
-	public static function GetFileSize(int $bytes, int $decimals = 2)
+	public static function GetFileSize(float $bytes, int $decimals = 2)
 	{
 		$factor = floor((strlen($bytes) - 1) / 3);
 		return sprintf("%.{$decimals}f", $bytes / pow(1024, $factor)) . self::UNIT_SIZES[$factor];
